finxed spelling mistake birdy to birdie throughout
