+++ source/core/src/main/com/csse3200/game/components/minigames/birdieDash/BirdieDashGame.java
@@ -1,0 +1,3 @@
+package com.csse3200.game.components.minigames.birdieDash;
+public class BirdieDashGame {
+}
