+++ source/core/src/main/com/csse3200/game/components/minigames/birdieDash/controller/BirdyDashController.java
@@ -1,0 +1,4 @@
+package com.csse3200.game.components.minigames.birdieDash.controller;
+public class BirdyDashController {
+    // To be implemented
+}
