+++ source/core/src/main/com/csse3200/game/components/minigames/birdieDash/entities/Bird.java
@@ -1,0 +1,4 @@
+package com.csse3200.game.components.minigames.birdieDash;
+public class Bird {
+    // To be implemented
+}
