+++ source/core/src/main/com/csse3200/game/components/minigames/birdieDash/entities/Coin.java
@@ -1,0 +1,4 @@
+package com.csse3200.game.components.minigames.birdieDash.entities;
+public class Coin {
+    // To be implemented
+}
