+++ source/core/src/main/com/csse3200/game/components/minigames/birdieDash/entities/Pipe.java
@@ -1,0 +1,4 @@
+package com.csse3200.game.components.minigames.birdieDash.entities;
+public class Pipe {
+    // To be implemented
+}
