+++ source/core/src/main/com/csse3200/game/components/minigames/birdieDash/rendering/BirdRenderer.java
@@ -1,0 +1,4 @@
+package com.csse3200.game.components.minigames.birdieDash.rendering;
+public class BirdRenderer {
+    // To be implemented
+}
