+++ source/core/src/main/com/csse3200/game/components/minigames/birdieDash/rendering/PipeRenderer.java
@@ -1,0 +1,4 @@
+package com.csse3200.game.components.minigames.birdieDash.rendering;
+public class PipeRenderer {
+    // To be implemented
+}
