--- source/core/src/main/com/csse3200/game/components/minigames/birdyDash/BirdyDashGame.java
@@ -1,3 +1,0 @@
-package com.csse3200.game.components.minigames.birdyDash;
-public class BirdyDashGame {
-}
