--- source/core/src/main/com/csse3200/game/components/minigames/birdyDash/controller/BirdyDashController.java
@@ -1,4 +1,0 @@
-package com.csse3200.game.components.minigames.birdyDash.controller;
-public class BirdyDashController {
-    // To be implemented
-}
