--- source/core/src/main/com/csse3200/game/components/minigames/birdyDash/entities/Bird.java
@@ -1,4 +1,0 @@
-package com.csse3200.game.components.minigames.birdyDash;
-public class Bird {
-    // To be implemented
-}
