--- source/core/src/main/com/csse3200/game/components/minigames/birdyDash/entities/Coin.java
@@ -1,4 +1,0 @@
-package com.csse3200.game.components.minigames.birdyDash.entities;
-public class Coin {
-    // To be implemented
-}
