--- source/core/src/main/com/csse3200/game/components/minigames/birdyDash/entities/Pipe.java
@@ -1,4 +1,0 @@
-package com.csse3200.game.components.minigames.birdyDash.entities;
-public class Pipe {
-    // To be implemented
-}
