--- source/core/src/main/com/csse3200/game/components/minigames/birdyDash/rendering/BirdRenderer.java
@@ -1,4 +1,0 @@
-package com.csse3200.game.components.minigames.birdyDash.rendering;
-public class BirdRenderer {
-    // To be implemented
-}
