--- source/core/src/main/com/csse3200/game/components/minigames/birdyDash/rendering/CoinRenderer.java
@@ -1,4 +1,0 @@
-package com.csse3200.game.components.minigames.birdyDash.rendering;
-public class CoinRenderer {
-    // To be implemented
-}
